+++ ext_emconf.php
@@ -4,7 +4,7 @@ $EM_CONF[$_EXTKEY] = array(
   'title' => 'Formhandler',
   'description' => 'The swiss army knife for all kinds of mailforms, completely new written using the MVC concept. Result: Flexibility, Flexibility, Flexibility  :-).',
   'category' => 'plugin',
-  'version' => '2.4.1e2',
+  'version' => '2.4.1e3',
   'state' => 'stable',
   'clearcacheonload' => true,
   'author' => 'Dev-Team Typoheads',
